+++ src/app/(components)/decrypt-modal/error-modal.tsx
@@ -1,53 +1,7 @@
 'use client'
 
-import { SubmitButton } from "@/components/ui/form"
-import { Icones, CloseIcon } from "@/components/ui/icons"
-import { useModal } from "@/components/ui/modal"
-import { ModalTitle } from "@/components/ui/modal/ModalUI"
-
-import wrapperBases from "@/components/ui/modal/index.module.css"
-import customStyles from "./error-modal.module.css"
 import Cipher from "@/components/ui/cipher"
-
-// ModalUIに相当するラッパー
-export function ErrorModalWrapper({ children }: {
-  children?: React.ReactNode,
-}) {
-  const { close } = useModal()
-
-  return (
-    <div className={wrapperBases['wrapper']} onClick={close}>
-      <div className={wrapperBases['content'] + ' ' + customStyles['content']} onClick={e => e.stopPropagation()}>
-        <button className={wrapperBases['close-button']} onClick={close}>
-          <Icones Icon={CloseIcon} fontSize={24} color="red-600" />
-        </button>
-        {children}
-      </div>
-    </div>
-  )
-}
-
-function ErrorModal({ children, title, message }: {
-  children?: React.ReactNode,
-  title: React.ReactNode
-  message: React.ReactNode
-}) {
-  const { close } = useModal()
-
-  return (
-    <div>
-      <ModalTitle className="text-red-600">
-        DECRYPTION ERROR!
-      </ModalTitle>
-      <div className="mb-5">
-        <h3 className="font-bold mb-1">[ {title} ]</h3>
-        <p>{message}</p>
-      </div>
-      {children}
-      <SubmitButton type="button" onClick={close} className={customStyles['submit-button']} autoFocus>OK</SubmitButton>
-    </div>
-  )
-}
+import { ErrorModal } from "@/components/ui/modal/error-modal"
 
 /** Base64エンコードでなかったときのモーダル */
 export function InvalidErrorModal({ privateKey }: { privateKey: string }) {
